+++ packages/terra-site/CHANGELOG.md
@@ -5,6 +5,9 @@ Unreleased
 ----------
 ### Added
 * Added new example to display date picker and popup inside modal.
+* Add test links for the Markdown component
+
+### Changed
 * Updated PropsTable rendering to include component title for packages that contain multiple components.
 
 1.2.0 - (July 26, 2017)